+++ packages/integration-sdk-runtime/src/execution/dependencyGraph.ts
@@ -94,7 +94,7 @@ export function executeStepDependencyGraph<
   function isStepEnabled(stepId: string) {
     return stepStartStates[stepId].disabled === false;
   }
-  StepResultStatus;
+
   /**
    * Updates the result of a step result with the provided satus
    */
@@ -171,7 +171,9 @@ export function executeStepDependencyGraph<
     typeTracker: TypeTracker,
   ) {
     const encounteredTypes = typeTracker.getEncounteredTypes();
-    const declaredTypesSet = new Set(getDeclaredTypesInStep(step));
+    const declaredTypesSet = new Set(
+      getDeclaredTypesInStep(step).declaredTypes,
+    );
     const undeclaredTypes = encounteredTypes.filter(
       (type) => !declaredTypesSet.has(type),
     );
@@ -346,11 +348,14 @@ function buildStepResultsMap<
             .dependenciesOf(step.id)
             .filter((id) => stepStartStates[id].disabled).length > 0;
 
+        const { declaredTypes, partialTypes } = getDeclaredTypesInStep(step);
+
         return {
           id: step.id,
           name: step.name,
           dependsOn: step.dependsOn,
-          declaredTypes: getDeclaredTypesInStep(step),
+          declaredTypes,
+          partialTypes,
           encounteredTypes: [],
           status:
             stepStartStates[step.id].disabled || hasDisabledDependencies
@@ -366,11 +371,21 @@ function buildStepResultsMap<
 
 function getDeclaredTypesInStep<
   TStepExecutionContext extends StepExecutionContext
->(step: Step<TStepExecutionContext>) {
-  const declaredTypes: string[] = [
-    ...step.entities.map((e) => e._type),
-    ...step.relationships.map((e) => e._type),
-  ];
+>(
+  step: Step<TStepExecutionContext>,
+): {
+  declaredTypes: string[];
+  partialTypes: string[];
+} {
+  const declaredTypes: string[] = [];
+  const partialTypes: string[] = [];
 
-  return declaredTypes;
+  [...step.entities, ...step.relationships].map((e) => {
+    declaredTypes.push(e._type);
+    if (e.partial) {
+      partialTypes.push(e._type);
+    }
+  });
+
+  return { declaredTypes, partialTypes };
 }